% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2_tax_table_cpp.R
\name{import_qiime2_tax_table_cpp}
\alias{import_qiime2_tax_table_cpp}
\title{Import QIIME2 Classification Table cpp}
\usage{
import_qiime2_tax_table_cpp(in_file)
}
\arguments{
\item{in_file}{A tab-delimited classification table output by QIIME2}
}
\value{
A phyloseq tax_table object
}
\description{
Import QIIME2 Classification Table cpp
}
\details{
This function expects up to 7 ranks (Domain, Phylum, Class, Order, Family, Genus and Species) but determines the number actually in the file.
}
\examples{
taxonomy_file <- read_tax_example("qiime2_tax_table.tsv")
example_tax_table <- import_qiime2_tax_table(in_file = taxonomy_file)
example_tax_table
\dontrun{
import_qiime2_tax_table_cpp(in_file = "qiime2_classifier_result.tsv")
}
}
