% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2_tax_table_cpp.R
\name{import_qiime2_taxtable_cpp}
\alias{import_qiime2_taxtable_cpp}
\title{Import QIIME Classification Table cpp}
\usage{
import_qiime2_taxtable_cpp(in_file)
}
\arguments{
\item{in_file}{A tab-delimited classification table output by QIIME2}
}
\value{
A phyloseq tax_table object
}
\description{
Import QIIME Classification Table cpp
}
\details{
This function expects up to 7 ranks (Domain, Phylum, Class, Order, Family, Genus and Species) but determines the number actually in the file.
}
\examples{
\dontrun{
import_qiime2_taxtable_cpp(in_file = "qiime2_classifier_result.tsv")
}
}
