% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rdp_tax_table_cpp.R
\name{import_rdp_tax_table_cpp}
\alias{import_rdp_tax_table_cpp}
\title{Import RDP Classifier Taxonomy Table cpp}
\usage{
import_rdp_tax_table_cpp(in_file, confidence)
}
\arguments{
\item{in_file}{A fix-rank tab-delimited text file output by the RDP Classifier}

\item{confidence}{The confidence level for filtering the taxonomy}
}
\value{
A phyloseq tax_table object
}
\description{
Import RDP Classifier Taxonomy Table cpp
}
\details{
A confidence value of 0.5 is recommended for shorter amplicons and a value of 0.8 for full-length 16S rRNA gene sequences.
}
\examples{
taxonomy_file <- read_tax_example("rdp_tax_table.tsv")
example_tax_table <- import_rdp_tax_table_cpp(in_file = taxonomy_file)
example_tax_table
\dontrun{
import_rdp_tax_table_cpp(in_file = "rdp_tax_table.tsv", confidence = 0.8)
}

}
