% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rescript_taxonomy.R
\name{import_rescript_taxonomy}
\alias{import_rescript_taxonomy}
\title{Import QIIME2 RESCRIPt Classification Table}
\usage{
import_rescript_taxonomy(in_file)
}
\arguments{
\item{in_file}{A tab-delimited classification table output by QIIME2 when one of my RESCRIPt classsifiers is used.for the classification.}
}
\value{
A phyloseq tax_table object
}
\description{
Import QIIME2 RESCRIPt Classification Table
}
\details{
This function is for use with QIIME2 taxonomy results when one of my classifiers built with RESCRIPt is used. These classifiers were built with the options p-rank-propagation and specifying the ranks Domain, Phylum, Class, Order, Family and Genus. However, a column for species is still expected (all entries are "s_"); it is removed by this function. I did not intend to include species because species assignments are usually nonsensical, especially with the SILVA database.
}
\examples{
\dontrun{
import_rescript_taxonomy(in_file = "qiime2_classifier_result.tsv")
}
}
