% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sintax_file_sorted.R
\name{import_sintax_file_sorted}
\alias{import_sintax_file_sorted}
\title{Import SINTAX Classifier Taxonomy Sorted}
\usage{
import_sintax_file_sorted(in_file, confidence)
}
\arguments{
\item{in_file}{A fix-rank tab-delimited text file output by SINTAX}

\item{confidence}{The confidence level for filtering the taxonomy}
}
\value{
A phyloseq tax_table object
}
\description{
Import SINTAX Classifier Taxonomy Sorted
}
\details{
A confidence value of 0.5 is recommended for shorter amplicons and a value of 0.8 for full-length 16S rRNA gene sequences.
}
\examples{
\dontrun{
import_sintax_file_sorted(in_file = "sintax_file.txt", confidence = 0.8)
}

}
