% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sintax_tax_table.R
\name{import_sintax_tax_table}
\alias{import_sintax_tax_table}
\title{Import SINTAX Taxonomy Table}
\usage{
import_sintax_tax_table(in_file, confidence)
}
\arguments{
\item{in_file}{A fix-rank tab-delimited text file output by SINTAX}

\item{confidence}{The confidence level for filtering the taxonomy (0.8 by default)}
}
\value{
A phyloseq tax_table object
}
\description{
Import SINTAX Taxonomy Table
}
\details{
A confidence value of 0.5 is recommended for shorter amplicons and a value of 0.8 for full-length 16S rRNA gene sequences.
}
\examples{
\dontrun{
import_sintax_tax_table(in_file = "sintax_tax_table.txt", confidence = 0.8)
}

}
