% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sintax_tax_table_cpp.R
\name{import_sintax_tax_table_cpp}
\alias{import_sintax_tax_table_cpp}
\title{Import SINTAX Taxonomy Table cpp}
\usage{
import_sintax_tax_table_cpp(in_file, confidence)
}
\arguments{
\item{in_file}{A fix-rank tab-delimited text file output by SINTAX}

\item{confidence}{The confidence level for filtering the taxonomy (0.8 by default)}
}
\value{
A phyloseq tax_table object
}
\description{
Import SINTAX Taxonomy Table cpp
}
\details{
A confidence value of 0.5 is recommended for shorter amplicons and a value of 0.8 for full-length 16S rRNA gene sequences.
}
\examples{
taxonomy_file <- read_tax_example("sintax_tax_table.tsv")
example_tax_table <- import_sintax_tax_table_cpp(in_file = taxonomy_file)
example_tax_table
\dontrun{
import_sintax_tax_table_cpp(in_file = "sintax_tax_table.tsv", confidence = 0.8)
}

}
