% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tax_example.R
\name{read_tax_example}
\alias{read_tax_example}
\title{Get path to example taxonomy files}
\usage{
read_tax_example(file = NULL)
}
\arguments{
\item{file}{Name of file to return.}
}
\value{
Returns the full path to the requested file from the package sub-directory inst/extdata if it exists, or a list of the files in the directory if no file is specified.
}
\description{
Access the files in speedyseq's inst/extdata sub-directory
}
\examples{
read_tax_example()
read_tax_example("rdp_able.tsv")
}
