% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_cap.R
\name{simple_cap}
\alias{simple_cap}
\title{Simple Capitalization}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{A string or vector of strings}
}
\value{
The string with every word capitalized
}
\description{
Simple Capitalization
}
\examples{
simple_cap("an old dog learning new tricks")

}
