% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedytax-package.R
\docType{package}
\name{speedytax-package}
\alias{speedytax}
\alias{speedytax-package}
\title{speedytax: Rapidly Import Classifier Results into 'phyloseq'}
\description{
Import classification results from the 'RDP Classifier' (Ribosomal Database Project),' 'USEARCH sintax,' 'vsearch sintax' and the 'QIIME2' (Quantitative Insights into Microbial Ecology) classifiers into 'phyloseq' tax_table objects.
}
\author{
\strong{Maintainer}: John Quensen \email{quensenj@msu.edu} (\href{https://orcid.org/0000-0002-7669-1595}{ORCID}) [copyright holder]

}
\keyword{internal}
