% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2_tax_table.R
\name{import_qiime2_tax_table}
\alias{import_qiime2_tax_table}
\title{Import QIIME2 Classification Table}
\usage{
import_qiime2_tax_table(in_file)
}
\arguments{
\item{in_file}{A tab-delimited classification table output by QIIME2}
}
\value{
A phyloseq tax_table object
}
\description{
Import QIIME2 Classification Table
}
\details{
This function expects 7 ranks: Domain, Phylum, Class, Order, Family, Genus and Species.
}
\examples{
\dontrun{
import_qiime2_tax_table(in_file = "qiime2_classifier_result.tsv")
}
}
