% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2_tax_table.R
\name{import_qiime2_tax_table}
\alias{import_qiime2_tax_table}
\title{Import QIIME2 Classification Table}
\usage{
import_qiime2_tax_table(in_file)
}
\arguments{
\item{in_file}{A tab-delimited classification table output by QIIME2}
}
\value{
A phyloseq tax_table object
}
\description{
Import QIIME2 Classification Table
}
\details{
This function expects 7 ranks: Domain, Phylum, Class, Order, Family, Genus and Species.
}
\examples{
taxonomy_file <- read_tax_example("qiime2_classification_table.tsv")
example_tax_table <- import_qiime2_tax_table(in_file = taxonomy_file)
example_tax_table
\dontrun{
import_qiime2_taxonomy(in_file = "qiime2_class_table.tsv")
}
}
