% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tax_example.R
\name{read_tax_example}
\alias{read_tax_example}
\title{Get path to example taxonomy files}
\usage{
read_tax_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
speedytax comes bundled with example taxonomy files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
read_tax_example()
read_tax_example("challenge.csv")
}
