% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tax_example.R
\name{read_tax_example}
\alias{read_tax_example}
\title{Get path to example taxonomy files}
\usage{
read_tax_example(file = NULL)
}
\arguments{
\item{file}{Name of file to return. If \code{NULL}, the files in inst/extdata will be listed.}
}
\value{
Returns the full path to the requested file from the package sub-directory inst/extdata if it exists. If called without specifying a file, it returns a list of files in inst/extdata. This function has the added benefit of working during package development when the system.file method cannot be used.
}
\description{
'speedytax' comes bundled with example taxonomy files in its \code{inst/extdata} directory. This function makes them easy to access.
}
\examples{
read_tax_example()
read_tax_example("rdp_classification_table.tsv")
}
