% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedytax.R
\docType{package}
\name{speedytax}
\alias{speedytax}
\alias{speedytax-package}
\title{speedytax}
\description{
Refactored functions for importing classifier results into phyloseq more quickly.
}
\author{
John Quensen \href{mailto:quensenj@msu.edu}{quensenj@msu.edu}
}
